Optimize StatusPagination
It is reported that id_in_range method is slow, so this will remove it.
It is intended to utilize an index of id, but that is possible by
removing created_at condition.
created_at condition is redundant because it is sorted by id, which is also
in the chronogical order.