Remove the flash notice after destroying OAuth authentication
The flash notice will not be shown before the user manually navigates to
certain pages, and the pages does not include the main Web UI. Therefore,
it cannot be expected that the user sees the message as soon as the
operation finishes.

The string just says "Successfully unlinked", and is likely to be
confusing if it is shown much later than the operation complete. This
change resolves the issue by removing the flash.