Do not test the user agent to change the link to pixiv
Apple required not to navigate to the sign up page, which prompts to input
private information when iOS app "Pawoo" was an application mainly
designed for Pawoo.

However, iOS app "Pawoo" is now a generic Mastodon client and is no longer
a client dedicated to Pawoo. Users are no longer prompted to use Pawoo
nor to input private information to pixiv. Therefore this code is no longer
necessary.